5730 Move DataFusionCommonConfig to its own file
